Move CTacticConfiguration traits
